visualization_step_2/fusion_visualization/requirements.txt
# Fusion Visualization Requirements
# System Requirements: Python 3.9.2+, PyTorch 2.0.0+, CUDA 11.7.0+

# Core PyTorch and Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torch-geometric>=2.3.0
# torch-scatter and torch-sparse require special installation - see notes below
# torch-scatter>=2.1.0
# torch-sparse>=0.6.17

# Scientific Computing
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning
scikit-learn>=1.0.0
timm>=0.4.12

# Visualization and Plotting
matplotlib>=3.4.0
seaborn>=0.11.0

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Explainable AI
grad-cam>=1.4.0
lime>=0.2.0
shap>=0.41.0

# Data Processing
tqdm>=4.60.0
PyYAML>=5.4.0
openpyxl>=3.0.0

# Additional dependencies for this module
networkx
h5py
dask[array]>=2021.6.0
dask-image>=0.6.0
pathlib

# Installation Notes:
# 1. Install PyTorch with CUDA 11.7+ support first:
#    pip install torch>=2.0.0 torchvision>=0.15.0 --index-url https://download.pytorch.org/whl/cu117
#
# 2. Install torch-scatter and torch-sparse (REQUIRED - must be installed separately):
#    pip install torch-scatter torch-sparse -f https://data.pyg.org/whl/torch-2.0.0+cu117.html
#
# 3. For complete pipeline dependencies, see main requirements.txt in project root 